fix: Added hook to load api google maps script and fix error. Linear issue: https://linear.app/quentral/issue/QUE-960/errori-in-console-google-maps
